almost finished webapplication -- frontend put, filters, search not finished

+++ KW33/GhibliMovieReviews/src/main/java/ch/noseryoung/GhibliMovieReviews/domain/movieReviews/MovieReviews.java
@@ -25,7 +25,7 @@ public class MovieReviews {
     @Column(name = "description")
     private String description;
     @Column(name = "length")
-    private Double lengthOfMovie;
+    private Integer lengthOfMovie;
     @Column(name = "image_url")
     private String image_url;
     @Column(name = "ghibli")
